docs: Add swagger OpenAPI docs to backend-v2. No endpoints described yet
Also: refactor routes to a separate folder
