+++ backend-v2/package.json
@@ -21,15 +21,19 @@
   "homepage": "https://github.com/strazdinsg/rebus#readme",
   "dependencies": {
     "@types/dotenv": "^6.1.1",
+    "@types/swagger-jsdoc": "^6.0.4",
+    "@types/swagger-ui-express": "^4.1.6",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
-    "mysql2": "^3.10.3"
+    "mysql2": "^3.10.3",
+    "swagger-jsdoc": "^6.2.8",
+    "swagger-ui-express": "^5.0.1"
   },
   "devDependencies": {
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
-    "@types/node": "^20.14.12",
+    "@types/node": "^20.14.13",
     "typescript": "^5.5.4"
   },
   "type": "module"
